add proper include file with wrappers
